+++ cmd/ethermintcli/main.go
@@ -1,139 +1,139 @@
-package main
+package test
 
-import (
-	"fmt"
+// import (
+// 	"fmt"
 
-	"github.com/spf13/cobra"
+// 	"github.com/spf13/cobra"
 
-	tmamino "github.com/tendermint/tendermint/crypto/encoding/amino"
-	"github.com/tendermint/tendermint/libs/cli"
+// 	tmamino "github.com/tendermint/tendermint/crypto/encoding/amino"
+// 	"github.com/tendermint/tendermint/libs/cli"
 
-	sdkclient "github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
-	clientrpc "github.com/cosmos/cosmos-sdk/client/rpc"
-	sdkcodec "github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/version"
-	"github.com/cosmos/cosmos-sdk/x/auth"
-	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	bankcmd "github.com/cosmos/cosmos-sdk/x/bank/client/cli"
+// 	sdkclient "github.com/cosmos/cosmos-sdk/client"
+// 	"github.com/cosmos/cosmos-sdk/client/flags"
+// 	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
+// 	clientrpc "github.com/cosmos/cosmos-sdk/client/rpc"
+// 	sdkcodec "github.com/cosmos/cosmos-sdk/codec"
+// 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+// 	sdk "github.com/cosmos/cosmos-sdk/types"
+// 	"github.com/cosmos/cosmos-sdk/version"
+// 	"github.com/cosmos/cosmos-sdk/x/auth"
+// 	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
+// 	"github.com/cosmos/cosmos-sdk/x/bank"
+// 	bankcmd "github.com/cosmos/cosmos-sdk/x/bank/client/cli"
 
-	"github.com/cosmos/ethermint/app"
-	"github.com/cosmos/ethermint/client"
-	"github.com/cosmos/ethermint/codec"
-	"github.com/cosmos/ethermint/crypto"
-	"github.com/cosmos/ethermint/rpc"
-	ethermint "github.com/cosmos/ethermint/types"
-)
+// 	"github.com/cosmos/ethermint/app"
+// 	"github.com/cosmos/ethermint/client"
+// 	"github.com/cosmos/ethermint/codec"
+// 	"github.com/cosmos/ethermint/crypto"
+// 	"github.com/cosmos/ethermint/rpc"
+// 	ethermint "github.com/cosmos/ethermint/types"
+// )
 
-var (
-	cdc = codec.MakeCodec(app.ModuleBasics)
-)
+// var (
+// 	cdc = codec.MakeCodec(app.ModuleBasics)
+// )
 
-func main() {
-	// Configure cobra to sort commands
-	cobra.EnableCommandSorting = false
+// func main() {
+// 	// Configure cobra to sort commands
+// 	cobra.EnableCommandSorting = false
 
-	tmamino.RegisterKeyType(crypto.PubKeySecp256k1{}, crypto.PubKeyAminoName)
-	tmamino.RegisterKeyType(crypto.PrivKeySecp256k1{}, crypto.PrivKeyAminoName)
+// 	tmamino.RegisterKeyType(crypto.PubKeySecp256k1{}, crypto.PubKeyAminoName)
+// 	tmamino.RegisterKeyType(crypto.PrivKeySecp256k1{}, crypto.PrivKeyAminoName)
 
-	keyring.CryptoCdc = cdc
-	clientkeys.KeysCdc = cdc
+// 	keyring.CryptoCdc = cdc
+// 	clientkeys.KeysCdc = cdc
 
-	// Read in the configuration file for the sdk
-	config := sdk.GetConfig()
-	ethermint.SetBech32Prefixes(config)
-	config.Seal()
+// 	// Read in the configuration file for the sdk
+// 	config := sdk.GetConfig()
+// 	ethermint.SetBech32Prefixes(config)
+// 	config.Seal()
 
-	rootCmd := &cobra.Command{
-		Use:   "ethermintcli",
-		Short: "Command line interface for interacting with ethermintd",
-	}
+// 	rootCmd := &cobra.Command{
+// 		Use:   "ethermintcd",
+// 		Short: "Command line interface for interacting with ethermintd",
+// 	}
 
-	// Add --chain-id to persistent flags and mark it required
-	rootCmd.PersistentFlags().String(flags.FlagChainID, "", "Chain ID of tendermint node")
-	rootCmd.PersistentPreRunE = func(_ *cobra.Command, _ []string) error {
-		return client.InitConfig(rootCmd)
-	}
+// 	// Add --chain-id to persistent flags and mark it required
+// 	rootCmd.PersistentFlags().String(flags.FlagChainID, "", "Chain ID of tendermint node")
+// 	rootCmd.PersistentPreRunE = func(_ *cobra.Command, _ []string) error {
+// 		return client.InitConfig(rootCmd)
+// 	}
 
-	// Construct Root Command
-	rootCmd.AddCommand(
-		clientrpc.StatusCommand(),
-		sdkclient.ConfigCmd(app.DefaultCLIHome),
-		queryCmd(cdc),
-		txCmd(cdc),
-		rpc.EmintServeCmd(cdc),
-		flags.LineBreak,
-		client.KeyCommands(),
-		flags.LineBreak,
-		version.Cmd,
-		flags.NewCompletionCmd(rootCmd, true),
-	)
+// 	// Construct Root Command
+// 	rootCmd.AddCommand(
+// 		clientrpc.StatusCommand(),
+// 		sdkclient.ConfigCmd(app.DefaultNodeHome),
+// 		queryCmd(cdc),
+// 		txCmd(cdc),
+// 		rpc.EmintServeCmd(cdc),
+// 		flags.LineBreak,
+// 		client.KeyCommands(),
+// 		flags.LineBreak,
+// 		version.Cmd,
+// 		flags.NewCompletionCmd(rootCmd, true),
+// 	)
 
-	// Add flags and prefix all env exposed with EM
-	executor := cli.PrepareMainCmd(rootCmd, "EM", app.DefaultCLIHome)
+// 	// Add flags and prefix all env exposed with EM
+// 	executor := cli.PrepareMainCmd(rootCmd, "EM", app.DefaultNodeHome)
 
-	err := executor.Execute()
-	if err != nil {
-		panic(fmt.Errorf("failed executing CLI command: %w", err))
-	}
-}
+// 	err := executor.Execute()
+// 	if err != nil {
+// 		panic(fmt.Errorf("failed executing CLI command: %w", err))
+// 	}
+// }
 
-func queryCmd(cdc *sdkcodec.Codec) *cobra.Command {
-	queryCmd := &cobra.Command{
-		Use:     "query",
-		Aliases: []string{"q"},
-		Short:   "Querying subcommands",
-	}
+// func queryCmd(cdc *sdkcodec.Codec) *cobra.Command {
+// 	queryCmd := &cobra.Command{
+// 		Use:     "query",
+// 		Aliases: []string{"q"},
+// 		Short:   "Querying subcommands",
+// 	}
 
-	queryCmd.AddCommand(
-		authcmd.GetAccountCmd(cdc),
-		flags.LineBreak,
-		authcmd.QueryTxsByEventsCmd(cdc),
-		authcmd.QueryTxCmd(cdc),
-		flags.LineBreak,
-	)
+// 	queryCmd.AddCommand(
+// 		authcmd.GetAccountCmd(cdc),
+// 		flags.LineBreak,
+// 		authcmd.QueryTxsByEventsCmd(cdc),
+// 		authcmd.QueryTxCmd(cdc),
+// 		flags.LineBreak,
+// 	)
 
-	// add modules' query commands
-	app.ModuleBasics.AddQueryCommands(queryCmd, cdc)
+// 	// add modules' query commands
+// 	app.ModuleBasics.AddQueryCommands(queryCmd, cdc)
 
-	return queryCmd
-}
+// 	return queryCmd
+// }
 
-func txCmd(cdc *sdkcodec.Codec) *cobra.Command {
-	txCmd := &cobra.Command{
-		Use:   "tx",
-		Short: "Transactions subcommands",
-	}
+// func txCmd(cdc *sdkcodec.Codec) *cobra.Command {
+// 	txCmd := &cobra.Command{
+// 		Use:   "tx",
+// 		Short: "Transactions subcommands",
+// 	}
 
-	txCmd.AddCommand(
-		bankcmd.SendTxCmd(cdc),
-		flags.LineBreak,
-		authcmd.GetSignCommand(cdc),
-		authcmd.GetMultiSignCommand(cdc),
-		flags.LineBreak,
-		authcmd.GetBroadcastCommand(cdc),
-		authcmd.GetEncodeCommand(cdc),
-		authcmd.GetDecodeCommand(cdc),
-		flags.LineBreak,
-	)
+// 	txCmd.AddCommand(
+// 		bankcmd.SendTxCmd(cdc),
+// 		flags.LineBreak,
+// 		authcmd.GetSignCommand(cdc),
+// 		authcmd.GetMultiSignCommand(cdc),
+// 		flags.LineBreak,
+// 		authcmd.GetBroadcastCommand(cdc),
+// 		authcmd.GetEncodeCommand(cdc),
+// 		authcmd.GetDecodeCommand(cdc),
+// 		flags.LineBreak,
+// 	)
 
-	// add modules' tx commands
-	app.ModuleBasics.AddTxCommands(txCmd, cdc)
+// 	// add modules' tx commands
+// 	app.ModuleBasics.AddTxCommands(txCmd, cdc)
 
-	// remove auth and bank commands as they're mounted under the root tx command
-	var cmdsToRemove []*cobra.Command
+// 	// remove auth and bank commands as they're mounted under the root tx command
+// 	var cmdsToRemove []*cobra.Command
 
-	for _, cmd := range txCmd.Commands() {
-		if cmd.Use == auth.ModuleName || cmd.Use == bank.ModuleName {
-			cmdsToRemove = append(cmdsToRemove, cmd)
-		}
-	}
+// 	for _, cmd := range txCmd.Commands() {
+// 		if cmd.Use == auth.ModuleName || cmd.Use == bank.ModuleName {
+// 			cmdsToRemove = append(cmdsToRemove, cmd)
+// 		}
+// 	}
 
-	txCmd.RemoveCommand(cmdsToRemove...)
+// 	txCmd.RemoveCommand(cmdsToRemove...)
 
-	return txCmd
-}
+// 	return txCmd
+// }
