+++ networks/local/ethermintnode/Dockerfile
@@ -23,7 +23,6 @@ RUN apt-get update
 
 # Copy over binaries from the build-env
 COPY --from=build-env /go/src/github.com/ChainSafe/ethermint/build/ethermintd /usr/bin/ethermintd
-COPY --from=build-env /go/src/github.com/ChainSafe/ethermint/build/ethermintcli /usr/bin/ethermintcli
 COPY --from=build-env /go/src/github.com/ChainSafe/ethermint/scripts/start.sh /
 
 EXPOSE 26656 26657 1317 8545 8546
